README: Link to an artifact archive, not individual files.
If you link to the browsable version of the artifact, it's a bit of a pain to
download the individual files. If we link to the archive itself, people will get
all the files they need in one go, which is much more convenient.

Also, if we're going to be linking to the artifact, let's give it a useful name,
and include the licence just to be safe.